Created the shareList form
